@@ -2,9 +2,9 @@ package com.iramml.uberclone.driverapp.Util;
 
 import android.annotation.SuppressLint;
 import android.content.pm.PackageManager;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+import androidx.appcompat.app.AppCompatActivity;
 import com.iramml.uberclone.driverapp.Interfaces.locationListener;
 import com.iramml.uberclone.driverapp.Messages.Message;
 import com.iramml.uberclone.driverapp.Messages.Messages;
