@@ -1,8 +1,8 @@
 package com.iramml.uberclone.driverapp.recyclerViewHistory;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -27,7 +27,7 @@ public class historyAdapter extends RecyclerView.Adapter<historyAdapter.ViewHold
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
-        View view= LayoutInflater.from(context).inflate(R.layout.history_template,viewGroup,false);
+        View view= LayoutInflater.from(context).inflate(R.layout.template_history,viewGroup,false);
         viewHolder=new ViewHolder(view, listener);
         return viewHolder;
     }
