@@ -2,8 +2,8 @@ package com.iramml.uberclone.riderapp.Activities;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -31,8 +31,6 @@ import com.google.android.gms.common.api.OptionalPendingResult;
 import java.util.Arrays;
 
 import mehdi.sakout.fancybuttons.FancyButton;
-import uk.co.chrisjenx.calligraphy.CalligraphyConfig;
-import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper;
 
 public class Login extends AppCompatActivity implements GoogleApiClient.OnConnectionFailedListener{
     private GoogleApiClient googleApiClient;
@@ -50,7 +48,6 @@ public class Login extends AppCompatActivity implements GoogleApiClient.OnConnec
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        CalligraphyConfig.initDefault(new CalligraphyConfig.Builder().setDefaultFontPath("fonts/NotoSans.ttf").setFontAttrId(R.attr.fontPath).build());
         setContentView(R.layout.activity_login);
         firebaseHelper=new FirebaseHelper(this);
         FancyButton signInButtonGoogle=findViewById(R.id.login_button_Google);
@@ -168,10 +165,5 @@ public class Login extends AppCompatActivity implements GoogleApiClient.OnConnec
     @Override
     public void onPointerCaptureChanged(boolean hasCapture) {
 
-    }
-
-    @Override
-    protected void attachBaseContext(Context newBase) {
-        super.attachBaseContext(CalligraphyContextWrapper.wrap(newBase));
     }
 }
