@@ -1,8 +1,8 @@
 package com.iramml.uberclone.riderapp.Activities;
 
 import android.app.AlertDialog;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -65,7 +65,7 @@ public class RateActivity extends AppCompatActivity {
     }
 
     private void submitRateDetails() {
-        final AlertDialog alertDialog=new SpotsDialog(this);
+        final AlertDialog alertDialog=new SpotsDialog.Builder().setContext(this).build();
         alertDialog.show();
 
         Rate rate=new Rate();
