@@ -3,7 +3,7 @@ package com.iramml.uberclone.riderapp.Util;
 import android.content.Context;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 
 import com.android.volley.AuthFailureError;
